added employee tests and changed file names
